Always output hex for <font color>
It makes sense to output stuff like <font color=blue> when the value is
actually provided by the author, maybe, but it causes interop problems
when the value is generated by the algorithms themselves, since the
format of computed styles is so poorly defined.  In particular, the
delete test "<p style=color:blue>foo<p style=color:brown>[]bar" was
generating <font color=brown> in Gecko and <font color=#a52a2a> in
WebKit for the *expected* result, and differences in *expected* results
are bad.

This behavior is simpler anyway.  It matches WebKit and Opera instead of
IE and Gecko.