Don't fail everything if execCommand() throws
Previously, just about everything would fail in IE, because it would
throw for styleWithCSS.  Of course, that was run as part of every test
except multitests.  As part of this change, I need to ignore changes to
the log div, because tests will be run in the interim and so the numbers
in the log div will naturally change.

Now IE notices all the tests and actually passes some, before they all
start failing with weird exceptions.